pyproject.toml
[tool.poetry]
name = "django_template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.4"
gunicorn = "20.1.0"
marshmallow_dataclass = "8.5.8"
pyyaml = "6.0"
psycopg2-binary = "2.9.1"
django-extensions = "^3.1.5"
ptpython = "^3.0.20"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
